requirements.txt
pandas
flask
flask_sqlalchemy
requests
pyyaml
